refactor: extract canvas preview types into leaf module
Move CANVAS_IMAGE_PREVIEW_WIDGET, CANVAS_IMAGE_PREVIEW_NODE_TYPES, and
supportsVirtualCanvasImagePreview into canvasImagePreviewTypes.ts — a
leaf module with only a type import. This eliminates the duplicated
inline constants in SubgraphNode and lets both SubgraphNode.ts and
promotionUtils.ts import from the same source without circular deps.

useNodeCanvasImagePreview.ts re-exports for backward compatibility.
